refactoring library extensions for flask app inside a common module
