--- services/users/project/utils/flaskcors.py
@@ -1,3 +1,0 @@
-from flask_cors import CORS
-
-cors = CORS()
