--- services/users/project/utils/ma.py
@@ -1,3 +1,0 @@
-from flask_marshmallow import Marshmallow
-
-ma = Marshmallow()
