--- services/users/project/utils/sqlalc.py
@@ -1,4 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-# instantiate the db
-db = SQLAlchemy()
