--- services/users/project/utils/toolbar.py
@@ -1,3 +1,0 @@
-from flask_debugtoolbar import DebugToolbarExtension
-
-toolbar = DebugToolbarExtension()
